+++ Changelog.md
@@ -6,6 +6,7 @@
 - Dropped PHP 8.3 support
 - Dropped Laravel 10 & 11 support 
 - Requires Vanilo 6
+- Added the channels and SKU filters to the product index 
 - Added Tax Category Selection to the Shipping Method Form
 - Added support for shipping line items
 - Added cards with proper info to the shipping method's show page